ImageMap/requirements.txt
numpy
opencv-python
torch
torchvision
ttkbootstrap
scikit-learn
pillow
babel
